DjangoVision/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
Deprecated==1.2.14
Django==5.0.1
gcloud==0.18.3
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.1
oauth2client==4.1.3
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
wrapt==1.16.0
